+++ nowcasting_dataset/data_sources/fake/__init__.py
@@ -1,0 +1,1 @@
+""" Make fake data functions """
