+++ tests/data_sources/gsp/test_gsp_model.py
@@ -5,7 +5,7 @@ import tempfile
 import numpy as np
 import pytest
 
-from nowcasting_dataset.data_sources.fake import gsp_fake
+from nowcasting_dataset.data_sources.fake.batch import gsp_fake
 from nowcasting_dataset.data_sources.gsp.gsp_model import GSP
